bump module version for 1.1.1 (v1104), update CHANGELOG.md, update upgrade-downgrade & country-updown scripts

+++ user/tests/scripts/country-updown/country-updown.sh
@@ -53,7 +53,9 @@ export PHOTON_SP2_070="https://github.com/particle-iot/device-os/releases/downlo
 export PHOTON_SP1_101="https://github.com/particle-iot/device-os/releases/download/v1.0.1/system-part1-1.0.1-photon.bin"
 export PHOTON_SP2_101="https://github.com/particle-iot/device-os/releases/download/v1.0.1/system-part2-1.0.1-photon.bin"
 export PHOTON_SP1_110="https://github.com/particle-iot/device-os/releases/download/v1.1.0/photon-system-part1@1.1.0.bin"
-export PHOTON_SP2_110="https://github.com/particle-iot/device-os/releases/download/v1.1.0/photon-system-part1@1.1.0.bin"
+export PHOTON_SP2_110="https://github.com/particle-iot/device-os/releases/download/v1.1.0/photon-system-part2@1.1.0.bin"
+export PHOTON_SP1_111="https://github.com/particle-iot/device-os/releases/download/v1.1.1/photon-system-part1@1.1.1.bin"
+export PHOTON_SP2_111="https://github.com/particle-iot/device-os/releases/download/v1.1.1/photon-system-part2@1.1.1.bin"
 
 
 # instead of these set options,
@@ -374,6 +376,13 @@ curl_all_required_system_parts() {
     curl --fail -L "$PHOTON_SP2_110" -o "system-part2-1.1.0-photon.bin"; if [[ "$?" -ne 0 ]]; then return 1; fi
   fi
 
+  if [ ! -e "system-part1-1.1.1-photon.bin" ]; then
+    curl --fail -L "$PHOTON_SP1_111" -o "system-part1-1.1.1-photon.bin"; if [[ "$?" -ne 0 ]]; then return 1; fi
+  fi
+  if [ ! -e "system-part2-1.1.1-photon.bin" ]; then
+    curl --fail -L "$PHOTON_SP2_111" -o "system-part2-1.1.1-photon.bin"; if [[ "$?" -ne 0 ]]; then return 1; fi
+  fi
+
   if [ ! -e "tinker-v0.4.9-photon.bin" ]; then
     curl --fail -L "$PHOTON_TKR_049" -o "tinker-v0.4.9-photon.bin"; if [[ "$?" -ne 0 ]]; then return 1; fi
   fi
@@ -604,14 +613,14 @@ fail() {
 #  87 |   JP2   |   050   | DEFAULT |   101      |   GB0   |   070/063/050   | wiped creds, re-enter
 #  88 |   JP2   |   050   | DEFAULT |   101      | DEFAULT |   070/063/050   | re-test, online
 # ----+---------+---------+---------+------------+---------+-----------------+----------------------
-#  89 |   JP2   |   050   | DEFAULT |   110      |   N/A   |   070/063/050   | online
-#  90 |   JP2   |   050   | DEFAULT | 051/110    |   N/A   |   070/063/050   | online
-#  91 |   JP2   |   050   | DEFAULT |   110      |   N/A   |   070/063/051   | online
-#  92 |   JP2   |   050   | DEFAULT |   110      |   JP2   |   070/063/050   | online
-#  93 |   JP2   |   050   | DEFAULT | 051/110    |   JP2   |   070/063/050   | online
-#  94 |   JP2   |   050   | DEFAULT |   110      |   JP2   |   070/063/051   | online
-#  95 |   JP2   |   050   | DEFAULT |   110      |   GB0   |   070/063/050   | wiped creds, re-enter
-#  96 |   JP2   |   050   | DEFAULT |   110      | DEFAULT |   070/063/050   | re-test, online
+#  89 |   JP2   |   050   | DEFAULT |   111      |   N/A   |   070/063/050   | online
+#  90 |   JP2   |   050   | DEFAULT | 051/111    |   N/A   |   070/063/050   | online
+#  91 |   JP2   |   050   | DEFAULT |   111      |   N/A   |   070/063/051   | online
+#  92 |   JP2   |   050   | DEFAULT |   111      |   JP2   |   070/063/050   | online
+#  93 |   JP2   |   050   | DEFAULT | 051/111    |   JP2   |   070/063/050   | online
+#  94 |   JP2   |   050   | DEFAULT |   111      |   JP2   |   070/063/051   | online
+#  95 |   JP2   |   050   | DEFAULT |   111      |   GB0   |   070/063/050   | wiped creds, re-enter
+#  96 |   JP2   |   050   | DEFAULT |   111      | DEFAULT |   070/063/050   | re-test, online
 # ----+---------+---------+---------+------------+---------+-----------------+----------------------
 # LAST TEST | ALL PASSED
 # -------------------------
@@ -3230,12 +3239,12 @@ fi
 
 if true; then
   # ------
-  # 110
+  # 111
   # ------
 
   heading
   echo "+----+---------+---------+---------+------------+---------+-----------------+------------------------+"
-  echo "| 89 |   JP2   |   050   | DEFAULT |   110      |   N/A   |   070/063/050   | online                 |"
+  echo "| 89 |   JP2   |   050   | DEFAULT |   111      |   N/A   |   070/063/050   | online                 |"
   echo "+----+---------+---------+---------+------------+---------+-----------------+------------------------+"
   enter_dfu_mode
   try set_country JP2
@@ -3247,10 +3256,10 @@ if true; then
   try exit_dfu_mode
 
   enter_dfu_mode
-  try dfu_system 1.1.0 photon
+  try dfu_system 1.1.1 photon
   try exit_dfu_mode
 
-  # wait for 1.1.0 bootloader from SMH
+  # wait for 1.1.1 bootloader from SMH
   sleep 30
 
   enter_dfu_mode
@@ -3273,7 +3282,7 @@ if true; then
 
   heading
   echo "+----+---------+---------+---------+------------+---------+-----------------+------------------------+"
-  echo "| 90 |   JP2   |   050   | DEFAULT | 051/110    |   N/A   |   070/063/050   | online                 |"
+  echo "| 90 |   JP2   |   050   | DEFAULT | 051/111    |   N/A   |   070/063/050   | online                 |"
   echo "+----+---------+---------+---------+------------+---------+-----------------+------------------------+"
   enter_dfu_mode
   try set_country JP2
@@ -3289,10 +3298,10 @@ if true; then
   try exit_dfu_mode
 
   enter_dfu_mode
-  try dfu_system 1.1.0 photon
+  try dfu_system 1.1.1 photon
   try exit_dfu_mode
 
-  # wait for 1.1.0 bootloader from SMH
+  # wait for 1.1.1 bootloader from SMH
   sleep 30
 
   enter_dfu_mode
@@ -3315,7 +3324,7 @@ if true; then
 
   heading
   echo "+----+---------+---------+---------+------------+---------+-----------------+------------------------+"
-  echo "| 91 |   JP2   |   050   | DEFAULT |   110      |   N/A   |   070/063/051   | online                 |"
+  echo "| 91 |   JP2   |   050   | DEFAULT |   111      |   N/A   |   070/063/051   | online                 |"
   echo "+----+---------+---------+---------+------------+---------+-----------------+------------------------+"
   enter_dfu_mode
   try set_country JP2
@@ -3327,10 +3336,10 @@ if true; then
   try exit_dfu_mode
 
   enter_dfu_mode
-  try dfu_system 1.1.0 photon
+  try dfu_system 1.1.1 photon
   try exit_dfu_mode
 
-  # wait for 1.1.0 bootloader from SMH
+  # wait for 1.1.1 bootloader from SMH
   sleep 30
 
   enter_dfu_mode
@@ -3353,7 +3362,7 @@ if true; then
 
   heading
   echo "+----+---------+---------+---------+------------+---------+-----------------+------------------------+"
-  echo "| 92 |   JP2   |   050   | DEFAULT |   110      |   JP2   |   070/063/050   | online                 |"
+  echo "| 92 |   JP2   |   050   | DEFAULT |   111      |   JP2   |   070/063/050   | online                 |"
   echo "+----+---------+---------+---------+------------+---------+-----------------+------------------------+"
   enter_dfu_mode
   try set_country JP2
@@ -3365,10 +3374,10 @@ if true; then
   try exit_dfu_mode
 
   enter_dfu_mode
-  try dfu_system 1.1.0 photon
+  try dfu_system 1.1.1 photon
   try exit_dfu_mode
 
-  # wait for 1.1.0 bootloader from SMH
+  # wait for 1.1.1 bootloader from SMH
   sleep 30
 
   enter_dfu_mode
@@ -3395,7 +3404,7 @@ if true; then
 
   heading
   echo "+----+---------+---------+---------+------------+---------+-----------------+------------------------+"
-  echo "| 93 |   JP2   |   050   | DEFAULT | 051/110    |   JP2   |   070/063/050   | online                 |"
+  echo "| 93 |   JP2   |   050   | DEFAULT | 051/111    |   JP2   |   070/063/050   | online                 |"
   echo "+----+---------+---------+---------+------------+---------+-----------------+------------------------+"
   enter_dfu_mode
   try set_country JP2
@@ -3411,10 +3420,10 @@ if true; then
   try exit_dfu_mode
 
   enter_dfu_mode
-  try dfu_system 1.1.0 photon
+  try dfu_system 1.1.1 photon
   try exit_dfu_mode
 
-  # wait for 1.1.0 bootloader from SMH
+  # wait for 1.1.1 bootloader from SMH
   sleep 30
 
   enter_dfu_mode
@@ -3441,7 +3450,7 @@ if true; then
 
   heading
   echo "+----+---------+---------+---------+------------+---------+-----------------+------------------------+"
-  echo "| 94 |   JP2   |   050   | DEFAULT |   110      |   JP2   |   070/063/051   | online                 |"
+  echo "| 94 |   JP2   |   050   | DEFAULT |   111      |   JP2   |   070/063/051   | online                 |"
   echo "+----+---------+---------+---------+------------+---------+-----------------+------------------------+"
   enter_dfu_mode
   try set_country JP2
@@ -3453,10 +3462,10 @@ if true; then
   try exit_dfu_mode
 
   enter_dfu_mode
-  try dfu_system 1.1.0 photon
+  try dfu_system 1.1.1 photon
   try exit_dfu_mode
 
-  # wait for 1.1.0 bootloader from SMH
+  # wait for 1.1.1 bootloader from SMH
   sleep 30
 
   enter_dfu_mode
@@ -3483,7 +3492,7 @@ if true; then
 
   heading
   echo "+----+---------+---------+---------+------------+---------+-----------------+------------------------+"
-  echo "| 95 |   JP2   |   050   | DEFAULT |   110      |   GB0   |   070/063/050   | wiped creds, re-enter  |"
+  echo "| 95 |   JP2   |   050   | DEFAULT |   111      |   GB0   |   070/063/050   | wiped creds, re-enter  |"
   echo "+----+---------+---------+---------+------------+---------+-----------------+------------------------+"
   enter_dfu_mode
   try set_country JP2
@@ -3495,10 +3504,10 @@ if true; then
   try exit_dfu_mode
 
   enter_dfu_mode
-  try dfu_system 1.1.0 photon
+  try dfu_system 1.1.1 photon
   try exit_dfu_mode
 
-  # wait for 1.1.0 bootloader from SMH
+  # wait for 1.1.1 bootloader from SMH
   sleep 30
 
   enter_dfu_mode
@@ -3525,7 +3534,7 @@ if true; then
 
   heading
   echo "+----+---------+---------+---------+------------+---------+-----------------+------------------------+"
-  echo "| 96 |   JP2   |   050   | DEFAULT |   110      | DEFAULT |   070/063/050   | re-test, online        |"
+  echo "| 96 |   JP2   |   050   | DEFAULT |   111      | DEFAULT |   070/063/050   | re-test, online        |"
   echo "+----+---------+---------+---------+------------+---------+-----------------+------------------------+"
 
   enter_ymodem
@@ -3541,10 +3550,10 @@ if true; then
   try exit_dfu_mode
 
   enter_dfu_mode
-  try dfu_system 1.1.0 photon
+  try dfu_system 1.1.1 photon
   try exit_dfu_mode
 
-  # wait for 1.1.0 bootloader from SMH
+  # wait for 1.1.1 bootloader from SMH
   sleep 30
 
   enter_dfu_mode
